requirements.txt
PyDispatcher==2.0.5
Scrapy==2.5.1
email_validator==1.1.1
phonenumbers==8.12.5
tldextract==2.2.2
Twisted==21.2.0
python-dotenv==0.19.2
tldextract==3.1.0